+++ templates/homepage/homepage.js
@@ -1,0 +1,1 @@
+/* Homepage related scripts go here */
